pipAR Prototype1
This is a simple prototype that tracks an image and places a transparent floor on  top of the existing floor and displays pipe locations underneath the ground. This is meant as an improvement on current GIS 2D tracking systems. The ultimate goal isto be able to transcribe 2D data into 3D space to make it easier to transcribe and make use of.